US3392 ENter Other Orgs
* removed new controller `GatewayLookupController` in favor of placing the
endpoint in the existing `GoVolunteerController`
