+++ Gateway/crds-angular.test/crds-angular.test.csproj
@@ -189,7 +189,6 @@
     <Compile Include="controllers\DonorControllerTest.cs" />
     <Compile Include="controllers\EventLocationControllerTest.cs" />
     <Compile Include="controllers\AccountControllerTest.cs" />
-    <Compile Include="controllers\GatewayLookupControllerTest.cs" />
     <Compile Include="controllers\GoVolunteerControllerTest.cs" />
     <Compile Include="controllers\LoginControllerTest.cs" />
     <Compile Include="controllers\LookupControllerTest.cs" />
