+++ Gateway/crds-angular/Controllers/API/GoVolunteerController.cs
@@ -1,8 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Web.Http;
 using System.Web.Http.Description;
 using crds_angular.Exceptions.Models;
 using crds_angular.Models.Crossroads.GoVolunteer;
+using crds_angular.Models.Crossroads.Lookups;
 using crds_angular.Security;
 using crds_angular.Services.Interfaces;
 
@@ -11,10 +13,12 @@ namespace crds_angular.Controllers.API
     public class GoVolunteerController : MPAuth
     {
         private readonly IOrganizationService _organizationService;
+        private readonly IGatewayLookupService _gatewayLookupService;
 
-        public GoVolunteerController(IOrganizationService organizationService)
+        public GoVolunteerController(IOrganizationService organizationService, IGatewayLookupService gatewayLookupService)
         {
             _organizationService = organizationService;
+            _gatewayLookupService = gatewayLookupService;
         }
 
         [HttpGet]
@@ -34,6 +38,23 @@ namespace crds_angular.Controllers.API
             catch (Exception e)
             {
                 var apiError = new ApiErrorDto("Get Organization failed: ", e);
+                throw new HttpResponseException(apiError.HttpResponseMessage);
+            }
+        }
+
+        [HttpGet]
+        [ResponseType(typeof(List<OtherOrganization>))]
+        [Route("api/organizations/other")]
+        public IHttpActionResult GetOtherOrganizations()
+        {
+            try
+            {
+                var Orgs = _gatewayLookupService.GetOtherOrgs();
+                return Ok(Orgs);
+            }
+            catch (Exception e)
+            {
+                var apiError = new ApiErrorDto("Unable to get other organizations", e);
                 throw new HttpResponseException(apiError.HttpResponseMessage);
             }
         }
